Simplify `move`
We thought that the `ds = [...]` was copied enough for it to warrant its
own function.

We collapsed a nested `if` into a single `and`.

It seemed more clear to unconditionally change the position and then
conditionally round than to conditionally change and round.

Well, I think it's clear, but David thinks it's just crafty, though
possibly less clear. :)